feat(right-manage): remove general file and separate panels
