+++ src/pages/ManageValidator/RightManagePanels/DepositPanel.tsx
@@ -3,13 +3,13 @@ import { YStack } from 'tamagui'
 import { useSelector } from 'react-redux'
 
 import PanelWrapper from './PanelWrapper'
-import { getDepositTitle } from '../../../../utilities'
-import { RootState } from '../../../../redux/store'
-import ConnectedWallet from '../../../../components/General/ConnectedWallet'
-import ConnectWallet from '../../../../components/General/ConnectWallet'
-import ChainParity from '../../../../components/General/ChainParity'
-import ValidatorRequest from '../../../../components/General/ValidatorRequest/ValidatorRequest'
-import ValidatorBalanceInfoBox from '../../../../components/General/ValidatorBalanceInfoBox'
+import { getDepositTitle } from '../../../utilities'
+import { RootState } from '../../../redux/store'
+import ConnectedWallet from '../../../components/General/ConnectedWallet'
+import ConnectWallet from '../../../components/General/ConnectWallet'
+import ChainParity from '../../../components/General/ChainParity'
+import ValidatorRequest from '../../../components/General/ValidatorRequest/ValidatorRequest'
+import ValidatorBalanceInfoBox from '../../../components/General/ValidatorBalanceInfoBox'
 import DepositValidator from './DepositValidator'
 
 const DepositPanel = () => {
