+++ src/pages/ManageValidator/RightManagePanels/ExitPanel.tsx
